+++ src/som/primitives/ActivitySpawn.java
@@ -19,8 +19,8 @@ import som.primitives.arrays.ToArgumentsArrayNodeFactory;
 import som.primitives.processes.ChannelPrimitives;
 import som.primitives.processes.ChannelPrimitives.Process;
 import som.primitives.processes.ChannelPrimitives.TracingProcess;
-import som.primitives.threading.TaskPrimitives.SomForkJoinTask;
-import som.primitives.threading.TaskPrimitives.TracedForkJoinTask;
+import som.primitives.threading.TaskThreads.SomForkJoinTask;
+import som.primitives.threading.TaskThreads.TracedForkJoinTask;
 import som.primitives.threading.ThreadPrimitives.SomThread;
 import som.primitives.threading.ThreadingModule;
 import som.vm.VmSettings;
